# [WIP] Virtual Tags for Discriminated Unions