Discriminated Unions are represented internally as a shallow object hierarchy. Externally only the base class is exposed. Determining which actually type it is (i.e. some for of RTTI) is broadly handled using one of two mechanisms, the mechasim being determined by how many union cases exist (there are some optimizatins around using NULL, and when Nullary items exists).

The first mechanism uses .net RTTI for determining underlying cases, and the second uses a Tag field stored in the base class, which is populated by by the case on construction. When four or more cases exist, the second method is used.

The first method is lighter on memory, but somewhat slower, especially when threes cases exist, the second method is faster, but consumes an extra 4 bytes per object instance.

The PR introduces a third, opt-in method, which is a halfway house between the two methods.  It creates an abstract Tag on the base class, and each of the derived types just return a constant value.

Speed wise this falls between the two existing methods, and memory wise it is as good as the first.

Opt is has been handled by adding another flag to CompilationRepresentationFlags (currently it overrides UseNullAsTrueValue - as in ORing them together doesn't work, although it could be used in conjunction, which would be faster for the null case;)

I'll create some performance number soon. 

This is represented as a WIP to determine if this is a) a good idea, and b) using CompilationRepresentationFlags is a good idea. Also, I've only really just done this as a proof of concept really so I don't know if I'm broken things at the moment...