@manofstick 2016-09-04
OK - some numbers!

Not that it is probably practical, because it would cause all sorts of compatibility problems (maybe?), but one area that could benefit from such a change would be the humble Map.

So if we change the UseNullAsTrueValue to UseVirtualFlag in MapTree and then run some tests...

    [<CompilationRepresentation(CompilationRepresentationFlags.UseVirtualTag)>]
    [<NoEquality; NoComparison>]
    type MapTree<'Key,'Value when 'Key : comparison > = 
        | MapEmpty 
        | MapOne of 'Key * 'Value
        | MapNode of 'Key * 'Value * MapTree<'Key,'Value> *  MapTree<'Key,'Value> * int
            // REVIEW: performance rumour has it that the data held in MapNode and MapOne should be
            // exactly one cache line. It is currently ~7 and 4 words respectively. 


So lets [hammer the map](https://gist.github.com/manofstick/08cc49dbd1b8d0d91571e2630e6462b6) with some data.

 | Previous Time (ms) | Modified Time (ms) | % of Original
------------ | ------------- | ------------- | -------------
x86 build | 1622 | 1395 | 86%
x86 lookup | 413 | 373 | 90%
x64 build | 1758 | 1523 | 87%
x64 lookup | 404 | 384 | 95%

So it's about ~ 10% faster. Pretty good?