@manofstick 2016-09-05
- the build errors, as far as I'm aware, are just because I have increased the public surface area with the new flag; I'm not sure if I should me modifying such things though (i.e. shouldn't change surface without someone with property authority I think?)
- in some contrived testing I found that having the IntegerTag field was fastest, but when I forced IntegerTag encoding with the Map example as per the timing from the previous message I actually found negligible time difference. Thus I think that the VirtualTag is superior to both RuntimeTypes (better speed, equal space) and IntegerTag (better space, equal speed).