@forki 2016-09-05
> just because I have increased the public surface area

you should still change that test then. review on green is better