@manofstick 2016-09-06
@forki 

OK; well we'll see how she goes; my machine just doesn't run the test suite and I haven't bothered to find out why, and NUnit integration in Visual Studio doesn't work either, so I just ramshackled some stuff together to make the surface area check. Should have just manually put it in, would have been easier and quicker, but hey. (don't have much time to give this machine any love; just happy it compiles!)

But I guess I would have liked to hear some feedback, not particularly about my implementation, as that is neither here not there, but rather about the idea. Because to me this method seems like the way to go when the existing IntegerTag and RunTimeTypes methods could be deprecated?  This could simplify things vastly, but I'm not sure of the ramifications of such a rollout, due to inlining especially which would mean version management? But maybe that's already an issue with any FSharp.Core change?

(Not that this matters in my currently implementation, as it's opt-in, but I would like to see some discussion in the general case....)