@dsyme 2016-09-06
@manofstick The implementation look good and seems a pretty sane design.  The performance improvements are convincing.  Would like to see if overall compiler perf is improved by this (see also Zmap in the compiler)