@manofstick 2016-09-07
@dsyme 

Zmap (and Zset) are optimized for performance already; i.e. they are only a single nullary/single non-nullary case, and they include UseNullAsTrueValue, so they optimize down to just null checks - although that does mean from a space perspective they are a little more bloated as the don't contain the (MapOne|SetOne) case. (Must have been experimental at some stage, as it's wrapped in a conditional compiler directive "ONE")

I tried forcing all RuntimeTypes/IntegerType to VirtualTag, but FSharpOption was clashing with some members; I will have to look what combination causes issues. Otherwise I might just sprinkle it the flag around - do you have a list of most used types from a performance tests? Do you have any standard performance tests for the compiler?

And, if this was to be come non-opt-in (i.e. default for all disciminate unions - although keep the special cases for all nullary and single non-nullary/single nullary and structs) would this cause compatibility issues; i.e. with inlining of modules that were build pre-such a change? Or is this not an issue? I don't know; just trying to find out...