@dsyme 2016-09-08
> And, if this was to be come non-opt-in ... would this cause compatibility issues

 I think the biggest concern is that serialization format incompatibility (e.g. binary serialization) .  Inlining etc. would be ok.