@manofstick 2016-09-08
@dsyme 

Would inlining be OK? If I have an assembly built with a previous version, and then tried to reference it in an assembly built with a new one, then the new assembly, in say a match expression would try to make explicit access to the underlying virtual Tag member, which just wouldn't exist? Or is this not how it would work?

Anyway, hadn't thought of binary serialization. Bummer. I guess that really does mean it will have to remain opt-in. A shame. Because RuntimeTypes is always worse, and I'm pretty sure the difference in the real world between the VirtualTag and IntegerTag is fairly negligible performance wise (caveat; I really need to do some better testing!) and you get your 4 bytes per object saving... Hmmm...