@manofstick 2016-09-08
@forki 

I saw you had [FSharpPerf](https://github.com/fsprojects/FSharpPerf) project, which sounds just like what I need, but I tried to do "build.cmd BuildProjects" with the default build.fsx but it failed; I pressed on anyway with the RunPerfTests, but it just failed too. Should I not be using defaults? Is there anything else I need? Or do I really need to modify the build.fsx?