@forki 2016-09-08
from https://github.com/fsprojects/FSharpPerf#testing-on-appveyor

* Fork https://github.com/fsprojects/FSharpPerf
* Configure your F# compiler fork and the hashes to test in build.fsx
* Commit the change and send a PR