@manofstick 2016-09-08
@forki 

So just do a non-local build? OK.