@forki 2016-09-08
yeah. that's what I recommend. otherwise your machine is occupied for a while.