@forki 2016-09-08
https://github.com/fsprojects/FSharpPerf/pull/7 show cool results for this.