@manofstick 2016-09-08
@forki

Unfortunately I think these are the wrong way around! (Or at least that is what I would be expecting) I.e. I have probably slowed things down, as I have removed some exciting optimisations by just forcing everything to use virtual tag.

Anyway, I'll analyse and confirm, fix, and be more surgical now that I have a framework to work within.