@forki 2016-09-08
9d5e36a is your commit and FSharpPerf shows this as faster.