@manofstick 2016-09-10
@forki 

Ahh but that commit was the baseline; the other one was the one where I forced all union types to use virtual tag! No great performance leap!

And I'm not really expecting much, if any, improvement in regards to the compiler, as I believe that people have manually optimized the cases where things were slow by cutting down the number of cases to a single non-nullary/single nullary cases such as what was done with Zmap/Zset. These end up as best case scenerios, as they only need to do null checks with no tags anyway. And in my original test I converted these to my less optimal VirtualTag just because I wanted to see what would happen if I just did sweeping changes.

But what I would like to see is, if I can now pull out all those special cases and leave them in their optimized state, and then change the rest of the unions to VirtualTag I'm hoping this will have minimal slowdown or possibly even a slight performance increases then that would be a win; as we're then saving 4 bytes per union cases allocation.

Anyway, we'll see how we go - Ballet and [Mr Maker](http://www.mistermaker.com/) are providing me with a brief reprieve, but it won't last for long :-)